// backend_db/frontend/src/client/types.gen.ts
export type TableRead = {
    /**
     * Name
     */
    name: string;
    /**
     * Table Schema
     *
     * JSON schema definition for table structure
     */
    table_schema: {
        [key: string]: unknown;
    };
    /**
     * Public
     */
    public?: boolean;
    /**
     * Description
     */
    description?: string | null;
    /**
     * Metadata
     */
    metadata?: {
        [key: string]: unknown;
    } | null;
    /**
     * Id
     */
    id: string;
    /**
     * Owner Id
     */
    owner_id: string;
    /**
     * Row Count
     */
    row_count?: number;
    /**
     * Realtime Enabled
     */
    realtime_enabled?: boolean;
    /**
     * Created At
     */
    created_at?: string | null;
    /**
     * Updated At
     */
    updated_at?: string | null;
};